Add basic View capability, including diff
Provides a new View (F4) command that automatically opens the file in VS  Code. If the selected file is a modified file, opens a diff with the unmodified version.

Added global selection tracking, simplified main view (just use the commit view), and bump gui.cs to a newer version we're now building from a fork, since the upstream is lagging behind in nuget.org.

Fixes #2