Add command order inference based on the hotkey
It will be quite typical for commands to be ordered according to their hotkey, such as F1 followed by F2 and so on. So make the order optional and infer it from the enum value, since the values are already sorted that way.

NOTE: for key combinations (i.e. Alt + a key), you can still provide a specific order to override this default behavior. It's just a convenience.

Fixes #6