Added support for switching main view and commands
The new base class `MainView` encapsulates the command bar building, so that individual views can just pass along their imported global and local commands. The `context` for the local commands is used as the contract name for filtering.

This is work in progress since we need to move some types to a new `Guit.Plugin` nuget package which can be consumed by the various plugins.

Also, we're currently hardcoding the "main view" on app start, but that should also be contributed by plugins.

There is an example showing `Log` (with no contextual commands) and `Sync` (with `Fetch` and `Pull` contextual commands) to prove the concept.

Partially fixes #5.