Change JTF for MainThread
The JTF adds complexity that is not worth it for the target app types
and the typical behaviors we have. In fact, it can introduce hard to
diagnose bugs like using Run() when you should do RunAsync, on others.

We instead provide a very simple abstraction over the main loop with
MainThread, which allows to efficiently call Invoke (by first checking
if invoke is required at all, since it's not needed if you're already
on the main thread) and added the capability of retrieving a result
from that execution too (i.e. to get a dialog result).

This is sufficient and much simpler than the alternative.