Move plugin-specific dialogs to their respective plugin folder
These dialogs are not reusable outside of a specific plugin.
When we move the reusable and base dialogs to a "core" library,
the plugin-specific ones should remain in their respective
areas. At some point, those plugins will also get their own
assembly and nuget package, at which point we'll be able to
just move the entire folder to a new project with the right
project and package references.