Migrate to new dialog initialization support in gui.cs
This feature added to [gui.cs](https://github.com/migueldeicaza/gui.cs/pull/261)
allows dialogs to have a more intuitive and built-in supported pattern for building
the UI at a predictable point in time.

Update our dialogs to this new pattern, which simplifies quite a bit of the code
too.