Rename MainThread > ThreadContext and support background+main thread
For performance reasons, we have scenarios where we want to ensure code
is running on a background thread, as well as on the main thread. Therefore,
the `MainThread` name wasn't good enough anymore.

We add both `MainThread` and `Background` properties that allow using the
existing `Invoke/Invoke<T>` methods while keeping nice and short method
names (instead of `InvokeOnMainThread` or similarly ugly names).