Load Changes as a plugin
Added plugin discovery (for now only for built-in ones), automated
registration of built-in plugins into the main guit app by simply
adding a project reference to them (via `Directory.Build.*` on the
`Plugins` folder.

To test the selection mechanism, run the following on the git repo
root to test:

```
git config -e --local
```

In the editor that opens, add the following configuration section:

```
[guit]
	# Disable built-in plugins, so we can pick & choose
	coreplugins = false
	# Pick just the changes one
	plugin = Guit.Changes.dll
```

Run `guit` pointing to that directory and the changes view should be
the only one loaded. However, because the other two areas (Sync, Log)
have not been made into plugins yet, all of the commands will be
visible ¯\_(ツ)_/¯