Added ability to reload shell and plugins
The existing App has been renamed Shell, and App now is responsible
for restarting the Shell, after unloading and reloading all plugins.

Progress reporting of the pluging reloading is missing since I had
many issues attempting to refresh the UI with some progress since
the App is running in the main thread already, and has to continue
to do so in order to block the console program, even while restarting
the shell.

Optimized the plugin manager so that it intelligently detects the
need to perform a nuget restore (i.e. plugin version has changed)
or even call MSBuild at all (i.e. plugin assembly list points to
files that all exist on disk).