Added initial conflict resolution dialog
Launches VS if no merge tool is configured, but honors configured tool
if it uses `cmd` with replacement variables as documented in https://git-scm.com/docs/git-mergetool.

If the configured tool is `vsDiffMerge.exe` (without a full path), turns
it into a full path based on the same discovery and resolution as if no
tool was configured. This allows users to tweak the tool parameters, and still
get the improved resolution from the tool, in case `guit` is not launched
from a developer command prompt.

Missing ToString overriding for the conflict entries so they render the file
path.