Added support for cherry pick root mode
When there is no .releaseconfig file in the repository, the
cherry picker will be configured with the root repository
only.

And when the .releaseconfig file is present, the "Releaseator" mode
is used by mapping the submodules into the list of repositories.

Also added IDynamicMenuCommand to be able to show/execute commands
conditionally.

Added the SelectBranch (dynamic) command which it only available
in the root mode.