Make the command bar support and refresh dynamic commands
Involves hiding non-visible/enabled commands, which also allows
multiple commands to share the same hotkey (i.e. 'c' for both
Commit and Conflicts), where the first by order is the one that
ends up being rendered/executed.

In the Commit/Conflicts case, they are mutually exclusive, which
drives this use case to optimize the used space by sharing a
location and hotkey.

Implementing IRefreshPattern on the CommandBar is not working
properly so it has been commented. The commands are evaluated
but only on first run for now.