Produce guit package for even easier installation
Another common patern in dotnet tools is to name the package
as the command provided by the tool, in our case, `guit`. This
makes for a more concise install command too, but it may hinder
discoverability for users searching for `dotnet-` in nuget.org
to find useful tools, so we keep both for now.