Fix sorting of commands, assign help/select all highest order
We were not properly ordering by order (which was defaulted to 100) *and*
key, meaning that all views that did not provide a specific Order would
end up in whichever order MEF presented it. This manifested more clearly
in the new Help window, which drove the fix in all the relevant places.